Fix circular dependencies in codebase
- Extract base types (User, Source, Award, Ability, Action, etc.) to lib/types/base.ts
- Create lib/db/shared-converters.ts for converter functions without circular deps
- Move PaginateMonstersParams and similar types to service types.ts files
- Update imports to use types.ts directly instead of index.ts files
- Service converters now import from shared-converters instead of lib/db/converters

This reduces circular dependencies from 17 to 0.
