+++ lib/services/ancestries/repository.ts
@@ -16,7 +16,6 @@ import type { Source, User } from "@/lib/types";
 import type { CursorData } from "@/lib/utils/cursor";
 import { decodeCursor, encodeCursor } from "@/lib/utils/cursor";
 import { isValidUUID } from "@/lib/utils/validation";
-import type { PaginateAncestriesParams } from "./service";
 import type {
   Ancestry,
   AncestryAbility,
@@ -24,6 +23,7 @@ import type {
   AncestryRarity,
   AncestrySize,
   CreateAncestryInput,
+  PaginateAncestriesParams,
   SearchAncestriesParams,
   UpdateAncestryInput,
 } from "./types";
